refactor(azure): rename azureRepoUrl.ts to utilsAzure.ts
Names the module by what it is - the shared Azure DevOps helpers - rather
than by its single current function, so the next Azure helper the git
provider and the ticket connector both need has an obvious home.

Pure rename: no logic change. The header keeps the constraint that made
the module exist, so it does not drift back into importing sfdx-hardis
and closing the utils/index -> gitUtils -> ticketProvider/index cycle.
